requirements.txt
requests>=2.28.0
Pillow>=9.0.0
tkinter-tooltip>=2.0.0 